App refactoring
